+++ src/main/java/no/ntnu/stud/idatt2106/backend/repository/ExtraResidentTypeRepositoryImpl.java
@@ -32,7 +32,7 @@ public class ExtraResidentTypeRepositoryImpl implements ExtraResidentTypeReposit
   };
 
   @Override
-  public Optional<ExtraResidentType> findById(int id) {
+  public Optional<ExtraResidentType> findById(long id) {
     String sql = "SELECT * FROM extra_resident_type WHERE id = ?";
     return jdbc.query(sql, rowMapper, id).stream().findFirst();
   }
@@ -45,7 +45,7 @@ public class ExtraResidentTypeRepositoryImpl implements ExtraResidentTypeReposit
 
   @Override
   public void save(ExtraResidentType type) {
-    String sql = "INSERT INTO extra_resident_type (name, consumption_water, consumption_food) " 
+    String sql = "INSERT INTO extra_resident_type (name, consumption_water, consumption_food) "
         + "VALUES (?, ?, ?)";
     KeyHolder keyHolder = new GeneratedKeyHolder();
 
@@ -64,18 +64,18 @@ public class ExtraResidentTypeRepositoryImpl implements ExtraResidentTypeReposit
 
   @Override
   public void update(ExtraResidentType type) {
-    String sql = "UPDATE extra_resident_type SET name = ?, consumption_water = ?, " 
+    String sql = "UPDATE extra_resident_type SET name = ?, consumption_water = ?, "
         + "consumption_food = ? WHERE id = ?";
     jdbc.update(
-        sql, 
-        type.getName(), 
-        type.getConsumptionWater(), 
-        type.getConsumptionFood(), 
+        sql,
+        type.getName(),
+        type.getConsumptionWater(),
+        type.getConsumptionFood(),
         type.getId());
   }
 
   @Override
-  public void deleteById(int id) {
+  public void deleteById(long id) {
     jdbc.update("DELETE FROM extra_resident_type WHERE id = ?", id);
   }
 }
